requirements.txt
python-decouple
itertools
sklearn
tornado
pymongo